Set errors if a URL is not valid or cannot be persisted
This is a larger change that sets an error in a flash message when the
above events happen, and then retrieves them after being redirected to
the default route and renders them in the template. The intention is to
create something deliberately simplistic for the user, but log the
information internally all the same.