Update package.json
This change fleshes out the file so that it's more suitably descriptive
about the project, and about me as the author. That way, if someone reads
it, they're more likely to make sense of the project more easily, and be
able to get in touch with me, should they have issues.