Add Fly.io support
This is pretty new to me, but I think I've made enough progress to start
making meaningful deployments to the platform.

This change adds the Fly.io configuration file, fly.toml, adds a GitHub
workflow to deploy to Fly.io, .github/workflows/fly.yml, and updates the
Dockerfile to build in a way that models the Dockerfile generated by
flyctl launch, listed in the documentation.