Update the image configuration
After quite a lot of reading through the Fly.io docs and
experimentation, I think that there's no other way to provision the
database than to do so during container start up. A subsequent job in a
GitFlow workflow won't do it, as the app crashes because the database
isn't available. Fly's release commands don't work, as the volume isn't
available. My only thought, now, is to doing it as part of the container
starting, and hope that the volume is available at that time.