Cater to responsive resolutions
Up until now, I'd only given a token amount of consideration to
breakpoints smaller than a standard desktop. Given that, when viewed on
a mobile device the app would break in a number of places, such as the
results table.

This change corrects most of the issues with rendering the table at
smaller resolutions, but for the smallest three breakpoints uses a
div-based layout instead. Mainly because it just works better and is
easier to read.