Fix bug in rendering the table view
What I didn't see before, due to the test data, was that the table view
renders properly when the original URL cell has a large/long string in
it. Otherwise, the table breaks. This change, I believe, fixes this
issue so that the table renders properly despite the contents of a given
cell.